Passing AR and RANLIB to configure for szip/hdf5
HI,

Made some small modifications to cover the case where R was built with a compiler that is not in a standard location. The change introduces the variables AR and RANLIB to the configure steps for hdf5 and szip. This should not cause any issues regardless if R was built with LTO or not. Rhdf5lib compiles fine with or without -flto.

As you are probably aware, R 4.1.0 provides flags to control -flto when compiling packages. There is only one small case to cover after all this changes:

If you use --no-use-LTO, -flto is not set when compiling Rhdf5lib. Nevertheless, it is always set when building hdf5 and szip. I am not very sure how R controls this. So if R is built to use LTO, it will always compile hdf5 and szip with -flto. If there is someone using R with LTO support but not enabled unless you specify --use-LTO, szip and hdf5 will not be built with -flto.

Hope this makes sense and covers these edge cases.

Cheers.